+++ src/meson.build
@@ -15,10 +15,11 @@ endif
 
 subdir('lib')
 subdir('loaders')
+subdir('savers')
 subdir('bindings')
 
 thread_dep = meson.get_compiler('cpp').find_library('pthread')
-thorvg_lib_dep = [common_dep, loader_dep, binding_dep, thread_dep]
+thorvg_lib_dep = [common_dep, loader_dep, saver_dep, binding_dep, thread_dep]
 
 thorvg_lib = library(
 	'thorvg',
