+++ src/savers/tvg/meson.build
@@ -1,0 +1,9 @@
+source_file = [
+   'tvgTvgSaver.h',
+   'tvgTvgSaver.cpp',
+]
+
+subsaver_dep += [declare_dependency(
+    include_directories : include_directories('.'),
+    sources : source_file
+)]
